Add GetUser, GetMyMessages commands and test suite
- Add GetUser command to retrieve authenticated user profile
- Add GetMyMessages command to fetch eBay messages
- Add unit tests across all packages (commands, ebay, utils)
- Move shared types (Amount, BoolStr, Pagination) to structs.go
- Fix GetItemTransactions.Body() silently dropping NumberOfDays
- Fix GetTokenStatus redundant XMLName/Xmlns/Ack/Timestamp fields
- Fix ReviseFixedPriceItem.ItemSpecifics from unmarshalable map to *ItemSpecifics
- Normalize Quantity/QuantitySold to int for consistency

+++ pkg/commands/getItem.go
@@ -24,14 +24,14 @@ type GetItemResponse struct {
 	Item struct {
 		ItemID         string
 		Title          string
-		Quantity       int64
+		Quantity       int
 		PictureDetails *PictureDetails
 		ListingDetails struct {
 			StartTime string
 		}
 		SellingStatus struct {
 			ListingStatus string
-			QuantitySold  int64
+			QuantitySold  int
 			CurrentPrice  float64
 		}
 	}
